Initial commit
This has basic autocode working, even in cases with a nested directory
producing the autocode that gets used in the directory above. Not that
this doesnt use add_custom_target, but *does* seem to handle rebuilding
appropriately under both GNU make and Ninja. Next thing to get more
interesting would be to go to higher-order autocode; autocode that makes
inputs to other autocode and make sure that that functions properly.